Removed unsafe AsynchronousChannelGroup from tcp API
